# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Character": {
        "additionalProperties": false,
        "description": "DR-01, DR-02, DR-10. The whole of `cast/{id}/dossier.md`: identity, body, architecture.\n\nThe *wants / needs / lies* triple is not psychological decoration. `needs` usually\ncontradicts `wants` - the conscious goal is the wrong one - and that contradiction is\nwhat decides which way the character goes at each fork, which is precisely what the\nwriter has to resolve on the page. `lies` is the false belief about the self that keeps\nthe contradiction alive; the arc is the record of it giving way.\n\n`immutable_physical` is the anchor of invariant 3: these attributes do not change without\na registered ChangeEvent in `cast/{id}/changes.yaml`. FR-AGENT-06 checks the prose\nagainst this map, and finds a violation when a physical attribute in the draft differs\nfrom the value here with no ChangeEvent at or before the scene. Without the register, a\nscar that appears in chapter nine and a continuity error look identical on the page.\n\n**A dossier is never loaded whole.** FR-OPS-01 trims it to the instant of the scene: the\narc entry anchored at or before `story_time`, the knowledge rows acquired by then, the\nlatest valence of each relationship dated by then. The trim is not an optimisation. A\nwriter handed the complete record will use facts the character has not yet learned,\nbecause nothing in the text marks them as future - the record reads as true, and\neverything true in the context is fair to write.",
        "properties": {
          "arc": {
            "description": "Successive states, each anchored to a scene. Ordered by the story axis at read time, never assumed sorted on disk.",
            "items": {
              "$ref": "#/components/schemas/ArcEntry"
            },
            "title": "Arc",
            "type": "array"
          },
          "body": {
            "description": "The Markdown body: the character at the length a writer needs, for everything the structured fields above cannot carry.",
            "title": "Body",
            "type": "string"
          },
          "competences": {
            "description": "What the character can do. This bounds what she can solve: a problem outside this list has to be solved by somebody else or not at all.",
            "items": {
              "type": "string"
            },
            "title": "Competences",
            "type": "array"
          },
          "id": {
            "description": "Stable identifier; it names the `cast/{id}/` directory and keys the index row. The backend never renames (DR-08): renaming breaks every edge pointing here.",
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_-]*$",
            "title": "Id",
            "type": "string"
          },
          "immutable_physical": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Attributes that do not change without a registered ChangeEvent (invariant 3). A map because FR-AGENT-06 compares the prose against it key by key, and a ChangeEvent names the key it changed.",
            "title": "Immutable Physical",
            "type": "object"
          },
          "lies": {
            "description": "The false belief about the self that sustains the arc, and that the arc is the record of losing.",
            "title": "Lies",
            "type": "string"
          },
          "name": {
            "description": "What the character is called in the prose.",
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "needs": {
            "description": "What is actually missing. Usually contradicts `wants`, and that contradiction is what decides the choice at each fork.",
            "minLength": 1,
            "title": "Needs",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "wants": {
            "description": "The conscious goal pursued in scenes: what the character would say she is doing.",
            "minLength": 1,
            "title": "Wants",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "wants",
          "needs",
          "lies",
          "body"
        ],
        "title": "Character",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "The only process that reads and writes the harness stores.",
    "title": "My Story Marker \u2014 backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cast/{id}": {
      "get": {
        "description": "IF-03, `GET /cast/{id}`. The complete record.\n\n`GET /cast/{id}/dossier?at=` below is the same record trimmed to an instant (FR-OPS-01),\nand this one is its unsafe sibling: a writer handed the complete dossier uses facts the\ncharacter has not yet learned, because nothing in the text marks them as future. The\nrecord reads as true, and everything true in the context is fair to write. This route is\nfor a human reading their own tree, never for an assembled context.",
        "operationId": "read_character_cast__id__get",
        "parameters": [
          {
            "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "maxLength": 120,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9_-]*$",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Character"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a character's dossier",
        "tags": [
          "cast"
        ]
      }
    },
...
    "/manuscript/{id}": {
      "get": {
        "description": "IF-03, `GET /manuscript/{id}`. The prose of one scene, whole.\n\nThis route is for a reader -- a human, the frontend, the export -- and not for context\nassembly: raw draft prose never enters a model's context except as the previous scene's\n`literal_tail` (FR-OPS-03). A route that returns the manuscript is not a route that puts\nit in front of an agent.",
        "operationId": "read_draft_manuscript__id__get",
        "parameters": [
          {
            "description": "Stable scene id: `NNN`, three digits. Both files of a scene are named after it, at every digest level (DR-11).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable scene id: `NNN`, three digits. Both files of a scene are named after it, at every digest level (DR-11).",
              "pattern": "^\\d{3}$",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a draft",
        "tags": [
          "manuscript"
        ]
      },
      "put": {
        "description": "IF-04, `PUT /manuscript/{id}` (writer, style_editor).\n\nThe two roles that may write here reach it for different reasons -- the writer drafts,\nthe style editor polishes -- and the route does not tell them apart: which of them the\nrequest claims is `X-Agent-Role`'s answer, and whether that claim is allowed is Figure\n3's. Both are recorded on the provenance line, so a change of voice in a scene can be\ntraced to the role that made it.",
        "operationId": "write_draft_manuscript__id__put",
        "parameters": [
          {
            "description": "Stable scene id: `NNN`, three digits. Both files of a scene are named after it, at every digest level (DR-11).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable scene id: `NNN`, three digits. Both files of a scene are named after it, at every digest level (DR-11).",
              "pattern": "^\\d{3}$",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "description": "The Figure 3 role performing this write. Required on every write.",
            "in": "header",
            "name": "X-Agent-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The Figure 3 role performing this write. Required on every write.",
              "title": "X-Agent-Role"
            }
          },
          {
            "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Draft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvenanceRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write a draft",
        "tags": [
          "manuscript"
        ]
      }
    },
...
    "/scenes/{id}": {
      "get": {
        "description": "IF-03, `GET /scenes/{id}`. The whole record.\n\nEverything the harness does to produce a chapter is a function of this record: selection\nreads it (FR-OPS-02), assembly is anchored to it (FR-OPS-03), and the mechanical\ninvariants run over its `story_time`, `location`, `pov` and `participants`.",
        "operationId": "read_scene_scenes__id__get",
        "parameters": [
          {
            "description": "Stable scene id: `NNN`, three digits. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable scene id: `NNN`, three digits. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "pattern": "^\\d{3}$",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scene"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a scene record",
        "tags": [
          "scenes"
        ]
      },
      "put": {
        "description": "IF-04, `PUT /scenes/{id}` (architect).\n\nThe architect is a human acting under the role in v1 (spec Decision 5), which changes\nnothing here: the role still arrives in the header and Figure 3 still decides, because a\nhuman acts under a role rather than beside it.",
        "operationId": "write_scene_scenes__id__put",
        "parameters": [
          {
            "description": "Stable scene id: `NNN`, three digits. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable scene id: `NNN`, three digits. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "pattern": "^\\d{3}$",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "description": "The Figure 3 role performing this write. Required on every write.",
            "in": "header",
            "name": "X-Agent-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The Figure 3 role performing this write. Required on every write.",
              "title": "X-Agent-Role"
            }
          },
          {
            "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Scene"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvenanceRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write a scene record",
        "tags": [
          "scenes"
        ]
      }
    },
    "/scenes/{id}/assemble": {
      "post": {
        "description": "IF-05, `POST /scenes/{id}/assemble` (FR-OPS-03, FR-OPS-04, AC 12). Selects, then loads.\n\nThe selection is the one `/select` answers, made here first with its FR-IDX-08 update;\neach entry is then loaded as of the scene's story time, in ranking order after the fixed\nblock, the POV's dossier and the previous scene's tail, and the context is fitted to the\n100k cap: what did not fit is named in `removed` and `truncated_at`, what its as-of form\nexcludes in `withheld`, and nothing is cut inside an entry. `warnings` carries\n`fixed_block_over_budget` when the fixed block exceeds 800 tokens.\n\nAssembled with **no system prompt and no instruction**, so the estimate here is the\ndocuments alone. A turn's `dry_run` (plan step 18) assembles with the writer's system\nprompt and instruction counted in the mandatory part, and can therefore stop earlier than\nthis route on the same tree. A mandatory part over the cap on its own is a 422\n`ContextBudgetExceeded`: a record is too large and is fixed at the source (FR-CTX-05).\n\nNo `X-Agent-Role`: nothing in the stores is written. Plain `def` because it blocks on\nSQLite, the embedder and the store reads.",
        "operationId": "assemble_context_scenes__id__assemble_post",
        "parameters": [
          {
            "description": "Stable scene id: `NNN`, three digits. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable scene id: `NNN`, three digits. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "pattern": "^\\d{3}$",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssembledContext"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assemble the writer's context for a scene",
        "tags": [
          "scenes"
        ]
      }
    },
...
    "/scenes/{id}/select": {
      "post": {
        "description": "IF-05, `POST /scenes/{id}/select` (FR-OPS-02, AC 11). Ids, kinds and scores, never text.\n\nThe entities named in `pins` come first, in the record's order, then the axioms whose\n`scope` intersects `tags`, then BM25 and cosine fused by reciprocal rank; the POV is\nabsent. The FR-IDX-08 incremental update runs first, so the answer reflects every store\nwrite made before the call. A pin that names no entity is a 422 naming the `pins` field\nof the scene record, not a silent omission.\n\nNo `X-Agent-Role`: nothing in the stores is written, and the index is not a store (like\n`/index/rebuild`). Plain `def` because it blocks on SQLite and on the embedder.",
        "operationId": "select_entities_scenes__id__select_post",
        "parameters": [
          {
            "description": "Stable scene id: `NNN`, three digits. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable scene id: `NNN`, three digits. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "pattern": "^\\d{3}$",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Selection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select the entities a scene needs",
        "tags": [
          "scenes"
        ]
      }
    },
...
  }
}